+++ libs/compiler/codes.c
@@ -1446,7 +1446,7 @@ void tables_and_tree(const syntax *const sx, const char *const path)
 	*/
 
 	uni_printf(&io, "\n%s\n", "modetab");
-	for (size_t i = 0; (int)i < sx->md; i++)
+	for (size_t i = 0; i < sx->md; i++)
 	{
 		uni_printf(&io, "md %zi) %i\n", i, sx->modetab[i]);
 	}
